+++ src/Builder/LocalActivityBuilder.php
@@ -3,10 +3,10 @@
 namespace Keepsuit\LaravelTemporal\Builder;
 
 use InvalidArgumentException;
+use Keepsuit\LaravelTemporal\Facade\Temporal;
 use Temporal\Activity\LocalActivityOptions;
 use Temporal\Common\RetryOptions;
 use Temporal\Internal\Workflow\ActivityProxy;
-use Temporal\Workflow;
 use Temporal\Workflow\ActivityStubInterface;
 
 /**
@@ -39,12 +39,12 @@ final class LocalActivityBuilder
      */
     public function build(string $class)
     {
-        return Workflow::newActivityStub($class, $this->activityOptions);
+        return Temporal::getTemporalContext()->newActivityStub($class, $this->activityOptions);
     }
 
     public function buildUntyped(): ActivityStubInterface
     {
-        return Workflow::newUntypedActivityStub($this->activityOptions);
+        return Temporal::getTemporalContext()->newUntypedActivityStub($this->activityOptions);
     }
 
     public function __call(string $name, array $arguments): self